Update Dirac to lastest SVN
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1451 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/diracsplitter/libdirac/libdirac_byteio/dirac_byte_stats.cpp
@@ -1,6 +1,6 @@
 /* ***** BEGIN LICENSE BLOCK *****
 *
-* $Id: dirac_byte_stats.cpp,v 1.1 2006/04/20 10:41:56 asuraparaju Exp $ $Name: Dirac_0_9_1 $
+* $Id: dirac_byte_stats.cpp,v 1.3 2008/04/29 12:27:49 asuraparaju Exp $ $Name:  $
 *
 * Version: MPL 1.1/GPL 2.0/LGPL 2.1
 *
@@ -60,14 +60,14 @@ void DiracByteStats::Clear()
     m_byte_count.clear();
 }
 
-int DiracByteStats::GetBitCount(const StatType& stat_type) const  
+int64_t DiracByteStats::GetBitCount(const StatType& stat_type) const  
 {
     return GetByteCount(stat_type) * CHAR_BIT;
 }
 
-int DiracByteStats::GetByteCount(const StatType& stat_type) const
+int64_t DiracByteStats::GetByteCount(const StatType& stat_type) const
 {
-     std::map<StatType, int>::const_iterator it;
+     std::map<StatType, int64_t>::const_iterator it;
      it = m_byte_count.find(stat_type);
      if(it==m_byte_count.end())
            return 0;
@@ -75,7 +75,7 @@ int DiracByteStats::GetByteCount(const StatType& stat_type) const
       return it->second;
 }
 
-void DiracByteStats::SetByteCount(const StatType& stat_type, int count)
+void DiracByteStats::SetByteCount(const StatType& stat_type, int64_t count)
 {
     if(m_byte_count.find(stat_type)==m_byte_count.end())
         m_byte_count[stat_type]=0;
